pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "stonksbot"
version = "0.1.0"
description = "Stonksbot application"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "stonksbot"}]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.1.1"
aiogram = "^3.22.0"
pybit = "^5.7.0"
PyYAML = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.1.0"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
